gate the __dump commands behind -tags paritydump and mirror the w8 twins into test-scope
Claude Review round 2, both should-fixes. The __dump-* dispatch moves to
dump_gate.go behind -tags paritydump (make go-build-cli passes the tag,
go-lint vets both variants): __dump-foundations prints every configured
secret verbatim past the redaction twin — right for the parity gate,
wrong to compile into the artifact that becomes the shipped CLI at W19,
so the gate binary and the product binary are now different artifacts
and the leak cannot ship by omission; an untagged build answers the argv
with argparse's exit-2 invalid-choice error. contracts/cli/README.md
records the split. And _MIRRORED gains cli.py, changelog.py,
changelog_data.json and redaction.py — routed to the platform/security
areas, they scheduled neither the go nor the parity job, so a CLI flag
or a release's changelog append could merge green with the help/argv
goldens or the go:embed changelog copy stale (the same shape W7 closed
in 233cd73; paths/config/logging_setup are already GLOBAL).
